feat(fe2): component cleanup
