+++ packages/frontend-2/components/viewer/Controls.vue
@@ -81,7 +81,7 @@
     >
       <div v-show="activeControl === 'models'">
         <KeepAlive>
-          <ViewerResourcesList2
+          <ViewerResourcesList
             class="pointer-events-auto"
             @loaded-more="scrollControlsToBottom"
           />
