+++ packages/frontend-2/components/viewer/resources/ModelCard.vue
@@ -62,7 +62,7 @@
       v-show="showVersions && !showRemove"
       class="mt-2 ml-4 flex h-auto flex-col space-y-0"
     >
-      <ViewerResourcesVersionCard3
+      <ViewerResourcesVersionCard
         v-for="(version, index) in props.model.versions.items"
         :key="version.id"
         :model-id="modelId"
